restructured files and laid foundation for 2 layred cli
