+++ src/cli/main.java
@@ -1,0 +1,4 @@
+package cli;
+
+public class main {
+}
